+++ sites/www/changelog.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+- :bug:`-` `~paramiko.agent.AgentKey` had a dangling Python 3 incompatible
+  ``__str__`` method returning bytes. This method has been removed, allowing
+  the superclass' (`~paramiko.pkey.PKey`) method to run instead.
 - :release:`3.1.0 <2023-03-10>`
 - :feature:`2013` (solving :issue:`2009`, plus others) Add an explicit
   ``channel_timeout`` keyword argument to `paramiko.client.SSHClient.connect`,